requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-socketio>=5.10.0
pydantic>=2.5.0
httpx>=0.26.0
websockets>=12.0
pyyaml>=6.0
structlog>=24.1.0
